refactor(cli): use @depazer/server to provide api server
