+++ .changeset/healthy-tigers-try.md
@@ -1,0 +1,5 @@
+---
+'@depazer/cli': patch
+---
+
+♻️ Use @depazer/server to provide api server
