--- packages/cli/src/serve/index.ts
@@ -1,3 +1,0 @@
-import { startServer } from './startServer'
-
-export { startServer }
