--- packages/cli/test/server/router/module/api.spec.ts
@@ -1,8 +1,0 @@
-import apiRoute from '@/serve/router/module/api'
-import { describe, expect, test } from 'vitest'
-
-describe('apiRoute', () => {
-  test('should be an object', () => {
-    expect(typeof apiRoute).toBe('object')
-  })
-})
