--- packages/cli/test/server/router/module/base.spec.ts
@@ -1,8 +1,0 @@
-import baseRoute from '@/serve/router/module/base'
-import { describe, expect, test } from 'vitest'
-
-describe('baseRoute', () => {
-  test('should be an object', () => {
-    expect(typeof baseRoute).toBe('object')
-  })
-})
