--- packages/cli/test/server/router/staticServer.spec.ts
@@ -1,10 +1,0 @@
-import { handleStaticResource } from '@/serve/router/staticServer'
-import { ServerResponse } from 'http'
-import { describe, test, expect } from 'vitest'
-
-describe('handleStaticResource', () => {
-  test('resource not found', async () => {
-    const res = new ServerResponse({} as any)
-    expect(await handleStaticResource('/index.htm', res)).toBeNull()
-  })
-})
