--- packages/cli/test/server/startServer.spec.ts
@@ -1,8 +1,0 @@
-import { startServer } from '@/serve/startServer'
-import { test, describe } from 'vitest'
-
-describe('startServer', () => {
-  test('should start server', async () => {
-    await startServer(8080, 'test')
-  })
-})
